Extract shared file handling utilities into a new `file_utils` module
- Move commonly used functions (`get_file_path`, `is_audio`, `get_audio_extension`) and reusable constants (`SUPPORTED_AUDIO_TYPES`) from individual services to `file_utils`.
- Refactor `audio_handler`, `document_handler`, and `image_handler` to use the new module.
- Simplify and centralize MIME type handling for improved maintainability.
